requirements.txt
pillow
loguru
